Bug Report #12331
NVDA does not announce the names of the menus when walking with left and right arrows in iTunes. This can potentially be fixed by adding a class which helps accesses the text inside of an iTunes menu box. The function added below will attempt to provide that functionality. Further additions are above which add an object of this class to the clsList in chooseNVDAObjectOverlayClasses. A comment was not made on the bug report due to the issue unaffecting the user experience in any way and tight scheduling. Considering I have not contributed to NVDA before, this implementation may be going off the rails. However, I think I have provided a decent foundation in solving this issue.